refactor 검색모듈 패키지 구조 변경
